Remove hello.c and merge changes to fuse_fs.c (#8)
